Changes for 2.1.3.0
Changed Locator class to be a singleton to ensure reliability.

Refactored methods with "registerListener" of all types to be "addListener" for clarity and consistency.

Added convenience method to search for specific Tekdaqcs in the Locator class.

Added convenience method to add single channel listeners to analog and digital input objects.

Added alternate methods with default values for searching for Tekdaqcs in the Locator class.

Added ability to create "Unsafe" ATekdaqcs for the purpose of known static IP tekdaqcs on the network.

Made certain variables in LocatorResponse protected for creation of Unsafe ATekdaqcs